pyproject.toml
[tool.poetry]
name = "basic-mlops"
version = "0.1.0"
description = ""
authors = ["Maira <mairabaptista@hotmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
cookiecutter = "^1.7.3"
Cython = "^0.29.28"
scikit-learn = "^1.0.2"
Flask = "^2.1.2"
Flask-BasicAuth = "^0.2.0"
textblob = "^0.17.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
